feat: Replace CLI with a Slint-based GUI
This commit replaces the command-line interface of the file searcher with a graphical user interface built with Slint.

The new GUI provides the following features:
- A button to select a directory to index.
- A search bar to search for files based on their metadata.
- A list view to display the search results.
- A detail view to show the full metadata of a selected file.

The codebase has been refactored to separate the core logic (indexing, database) from the UI logic. The indexing process now runs in a separate thread to keep the UI responsive.
