+++ file-searcher/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.4.2", features = ["derive"] }
 walkdir = "2.3.3"
 rusqlite = { version = "0.29.0", features = ["bundled"] }
 kamadak-exif = "0.5.5"
@@ -15,3 +14,8 @@ serde_json = "1.0"
 anyhow = "1.0"
 humantime-serde = "1.1"
 mp4ameta = "0.13.0"
+slint = "1.5.1"
+rfd = "0.14.1"
+
+[build-dependencies]
+slint-build = "1.5.1"
