+++ file-searcher/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    slint_build::compile("src/ui/main.slint").unwrap();
+}
